+++ app/constants/Configuration.ts
@@ -1,3 +1,4 @@
 export const configuration = {
   MEETUP_API_BASE_URL: process.env.MEETUP_API_BASE_URL,
+  SETTINGS_BASE_URL: process.env.SETTINGS_BASE_URL,
 };