+++ app/types/domain/CodeCoffeeChapter.ts
@@ -1,0 +1,4 @@
+export type CodeCoffeeChapter = {
+  name: string;
+  meetupGroupUrlName: string;
+};